requirements.txt
streamlit
langchain
langchain-groq
pdfplumber
faiss-cpu
pylatexenc
python-dotenv
